Implemented most codegen (except var assignment)

+++ src/scanner.h
@@ -93,7 +93,7 @@ typedef struct token {
 	TokenType type;
 	TokenKeyword keyword;
 	char *data;
-	int integer;
+	long long int integer;
 	double number;
 	int line;
 } token_t;
